+++ main.py
@@ -64,6 +64,7 @@ def run():
     number_of_selected_units = 0
 
     # initialize the game
+    PLAYER_ID = 3
     size = "L"
     type_of_map = "island" # "mars_poles" "lake" "bridge" "island" "noise" "forest" "snow_forest"
 
@@ -81,10 +82,14 @@ def run():
     DICT_WITH_GAME_STATE = {
         "lowest_free_id": 1,
         "list_with_energy": [0, 10000, 10000, 10000, 10000],
+        "list_with_energy_last": [0, 0, 0, 0, 0],
+        "list_with_energy_current_production": [0, 0, 0, 0, 0],
         "list_with_energy_spent": [0, 0, 0, 0, 0],
-        "list_with_player_type": [False, "AI", "AI", "player", "AI"], # [False, "empty", "AI", "player", "empty"], [False, "empty", "empty", "player", "empty"], # 
+        "list_with_player_type": [False, "AI", "AI", "AI", "AI"], # [False, "empty", "AI", "player", "empty"], [False, "empty", "empty", "player", "empty"], # 
         "dict_with_new_units": {},
     }
+    DICT_WITH_GAME_STATE["list_with_player_type"][PLAYER_ID] = "player"
+
     DICT_WITH_UNITS = {}
     # make_test_units(DICT_WITH_GAME_STATE, DICT_WITH_UNITS)
     # LIST_WITH_UNITS = make_test_units_2()
@@ -144,6 +149,11 @@ def run():
             print(DICT_WITH_GAME_STATE["lowest_free_id"], end="\t")
             print()
 
+            # calculate energy
+            for i in range(1,5):
+                DICT_WITH_GAME_STATE["list_with_energy_current_production"][i] = DICT_WITH_GAME_STATE["list_with_energy"][i] - DICT_WITH_GAME_STATE["list_with_energy_last"][i]
+            DICT_WITH_GAME_STATE["list_with_energy_last"] = DICT_WITH_GAME_STATE["list_with_energy"].copy()
+
             # energy
             print("ENERGY:", end=" ")
             for player in range(1,5):
@@ -152,6 +162,10 @@ def run():
             print("SPENT:", end=" ")
             for player in range(1,5):
                 print(DICT_WITH_GAME_STATE["list_with_energy_spent"][player], end="\t")
+            print()
+            print("PROD.:", end=" ")
+            for player in range(1,5):
+                print(DICT_WITH_GAME_STATE["list_with_energy_current_production"][player], end="\t")
             print()
 
         for event in pygame.event.get():